Move the extra property row presenters/serializers to the Form namespace
The four classes translating between the definition form's builder rows
and the stored formats only exist for that form: they belong in
PrestaShop\PrestaShop\Core\ExtraProperty\Form next to
ExtraPropertyFormTypeMap and FormOptionsValidator, leaving Definition
and Validation to the domain-side classes (parser, value objects,
constraint mapper/catalog).
